# docs/swagger.yaml
basePath: /
definitions:
  finance.Message:
    properties:
      columns:
        items:
          type: string
        type: array
      data:
        items:
          items:
            type: string
          type: array
        type: array
      index:
        items:
          type: string
        type: array
    type: object
  types.Account:
    properties:
      id:
        type: string
    type: object
  types.FeedBody:
    properties:
      data:
        items:
          $ref: '#/definitions/types.FeedData'
        type: array
    type: object
  types.FeedData:
    properties:
      channel:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  types.HTTPError:
    type: object
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      operationId: get-string-by-int
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.HTTPError'
      summary: Show a account
  /rss-data:
    get:
      consumes:
      - application/json
      description: send mongodb data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.FeedBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.HTTPError'
      summary: Get mongodb data and send it back
  /tickers:
    get:
      consumes:
      - application/json
      description: get tickers in pandas format
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.HTTPError'
      summary: Get Yahoo stock tickers
swagger: "2.0"